rename: service 패키지의 파일들 위치 변경
