+++ src/main/java/bridge/PlayerSetting.java
@@ -5,8 +5,6 @@ import bridge.dto.BridgeCorrectLocationDto;
 import bridge.dto.BridgeDto;
 import bridge.dto.PlayerBridgeSizeDto;
 import bridge.exception.InputException;
-import bridge.BridgeMaker;
-import bridge.service.BridgeRandomNumberGenerator;
 
 import java.util.List;
