+++ src/main/java/bridge/domain/GameData.java
@@ -1,6 +1,6 @@
 package bridge.domain;
 
-import bridge.service.PlayerCommend;
+import bridge.PlayerCommend;
 
 import java.util.ArrayList;
 import java.util.List;
